+++ EdenSynth/libeden/source/synth/wavetable/Waveform.cpp
@@ -20,7 +20,7 @@ namespace eden::synth::wavetable
 	{
 	}
 
-	SampleType Waveform::operator()(float phase) const
+	SampleType Waveform::operator()(double phase) const
 	{
 		const int num2PI = static_cast<int>(phase / (2 * math_constants::PI));
 		phase -= num2PI * 2 * math_constants::PI;
